pyproject.toml
[tool.poetry]
name = "HIS-BE"
version = "0.1.0"
description = ""
authors = ["Doan Tien Loi <doantienloi123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
# be
python = "^3.10.9"
alembic = "1.11.3"
fastapi = "0.99.1"
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^12.1.1"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
sse-starlette = "1.6.5"
Unidecode = "^1.3.8"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
boto3 = "^1.34.131"
celery = "^5.2.0"
pymongo = "^4.8.0"
pusher = "^3.3.2"
redis = "^5.0.7"
flower = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
